requirements.txt
aniso8601==9.0.1
astroid==2.3.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
chart-studio==1.1.0
click==8.0.3
colorama==0.4.3
deprecated==1.2.13
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
isort==4.3.21
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.4.3
LivePopularTimes==1.3
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.5
pandas==1.3.5
plotly==5.5.0
pycparser==2.21
PyGithub==1.55
pyjwt==2.3.0
pylint==2.4.4
pynacl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.0
retrying==1.3.3
six==1.14.0
tenacity==8.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.11.2
llvmlite==0.38.0
timezonefinder==5.2.0
numba==0.55.0